sdk: update versions

+++ Linux/pipewire.sh
@@ -3,6 +3,9 @@
 source ./common.sh clang
 source ../common/clone-pipewire.sh
 
+if [[ -f $INSTALL_PREFIX/pipewire/bin/pipewire ]]; then
+  exit 0
+fi
 (
   cd pipewire-$PIPEWIRE_VERSION
   rm -rf build
